apps/primeins_intelligence/requirements.txt
streamlit
databricks-sdk>=0.40.0
psycopg2-binary
pandas
